Add page to create your own user realm (on same route as the user realm)
This is only shown if you are logged in, are allowed to create your own
user realm and are on your own user realm page which does not yet exist.
I felt like creating a separate route for that was not really useful.
Doing it right here saves the user one click.