Add shrinkability to search bar
This replaces `width: 100%` with `flex: 1` and declares
the input containing label as flex container.
This way, the search bar can actually shrink before the logo
when the screen is resized. Though that only becomes apparent with
a larger max-width (eg 600px or so) of the searchbar than we use now.