Fix size of language icon in header
Before this commit, it was always 24, regardless of screen size. So when
the other icons around it shrank, it didn't change size. I suspect its
size was set to a fixed "24" in response to one of my review comments
where I said it looked a bit big. Just removing this special case and
reconsidering that, I don't think its too big anymore. So just using
the same size for all icons seems sensible.