Set header's padding-left to 0 and adjust logos appropriately
Due to the recent header redesign, our default logo was shown a bit
small, which this commit fixes. This goes full in on "all logo margin is
in the SVG". Some background and additional information on this:

We allow universities to configure different logos. These logos have
different requirements (according to the university's CI) regarding
margin, i.e. blank space around the logo that must be present. In all
cases I've seen, that margin is relative to the logo size, e.g. the
ETH logo requires a margin of the height of the "E" in all directions.
And that seems to make sense: absolute margin doesn't really make sense.

We of course want to allow universities to abide by their CIs (how
generous of us). That's actually not really hard. Adding relative margin
is easy... it might waste a lot of space or render the logo smaller than
necessary. So our second goal is to not render too much superfluous
empty space and to force-shrink the logo as little as possible. That's
a lot harder.

This solution of "all margin must be in the SVG" puts more work on the
person creating said SVG, but it's likely the most straight-forward way
of letting people specify *relative margin*. If the logo is force-scaled
down when the header gets too narrow, the margin will shrink with the
logo, keeping the exact relative margin. Consider the alternative of
specifying some fixed margin in the config file, e.g. `left-margin: 16`.
Then when the logo is shrunk, the margin doesn't shrink and is thus too
much for the now smaller logo.

This solution is nice if all we care about is relative margin. However,
people might care about absolute margin as well. For example, one might
try to align the logo with the navigation element. In that case one
would want an absolute margin of 16px (as thats the margin-left of the
nav element). I think ideally, it would be possible to specify relative
and absolute margins in the config file. And Tobira would then take the
`max` of those two values. But I haven't found a good way to do that
AND it's probably overkill.

We might change the 16px of the nav element in the future anyway, so we
wouldn't want people to align to that. And a slight misalignment there
is really not a problem. I think it's really hard to notice.

Two other slight disadvantages of this approach:
- The focus outline of the logo is quite large now.
- The header height cannot shrink when the logo shrinks, leading to lots
  of whitespace above and below the header elements.