Make the header height decrease when the logo is scaled down
On narrow screens, with wide logos (like the ETH one), the logo gets
scaled down quite a bit due to lack of space in the x-axis. However,
the header height stayed the same, showing lots of empty space above
and below the logo. This commit addresses that by letting the header
shrink accordingly. Of course, there is a minimal height that the header
must have, given by its children + 2*16px of padding.

There are still some problems with the mobile version of the language,
user and burger menu. Their top element's height doesn't match the
header height anymore. And for the burger menu, the logo size jumps
(which I think happens before this PR already). An idea to solve this
would be to always render the normal header, but if those menus are
open, make all its children `opacity: 0` and render sth above it with
`position: absolute`.