Make "gallery view" of series block scale items to better fill screen
This should look better on more screens. Before, there was some
left-over space on the right if there wasn't enough space for a new
video item. On some screens it looked really bad because a new item
would almost fit.

With this, the items themselves scale a bit to fill the space. They do
have a max width. When that's reached, the remaining space is
distributed as margin left/right of the items.