Add `is_root` and `is_user_realm` fields to search_realm
That allows us to filter for those things during search. To not copy and
adjust the definition of `search_series`, they were adjusted in the
migration where they were created. We can do that as the migration is
not yet released. However, this brings up an ordering issue as the
search_realms view has to be adjusted first. So I reordered the
migrations.