Refactor and optimize insertion of dummy blocks
This adds a wrapper function which contains logic for recursively calling
an inner function to do operations on all realms.
The actual blocks are then added in two calls to this wrapper:
First, realms without any block get a matching series, if that exists.
Then, the remaining series are distributed over any other yet blockless realm,
along with additional text and video blocks.