Rewrite how we build filters for Meilisearch to fix bug
Previously the normal search would fail if being logged in as admin
due to a string building bug. Now I added a type to more semantically
represent filters, so that filter building is done first, and then it
is formatted as string. This likely leads to more heap allocations but
we had bugs in this code a couple times already, and correctness
obviously is most important here. I am not super happy with the helper
yet, but it can still be improved in the future.