Add breadcrumbs to all pages (excluding homepage)
This removes the logo tooltip and adds breadcrumbs to the `about tobira`,
`create user realm`, `page not found` and `search result` pages.
When there is no sidebox navigation, the breadcrumbs are aligned
like they would be on smaller screens, i.e. with the upper left corner of
the containing element.