Replace `pem` crate with `pem-rfc7468`
I changed this as the new crate name rolls off the tongue more easily.

Ehm, because we had `pem-rfc7468` in our dependency graph anyway as it
is used by other crypto dependencies. And it seems like it's
specifically designed for crypto use (constant time and stuff).

Potentially this means that Tobira is more strict on PEM formats now.
I think all PEM encoded keys actually used in practice are valid in the
strict format defined by RFC7468. But yeah, just in case people report
problems with their JWT keys later: this is likely the commit to blame.