Update `prometheus-client`
This was a bit of pain. API changed quite a bit and the new LabelSet
trait is weird. Also, there is a `Collector` trait now that seems to
be made for the "on demand" gathering of information we do. But
rewriting everything with that trait takes time and I don't see a
pressing need to do that.