Update PostgreSQL in CI (for checking migrations) from 10 to 11
See previous commit: 10 is EOL, so we don't need to check it. CI
wants to check with the oldest supported PG version.

Loading the DB dump unfortunately generates some warnings which lead to
a non-0 error code. But the import is fine, so we sadly have to ignore
the error code.