Increase `margin-bottom` of `h1` from 16 to 20
I saw that Bern university added an empty text block as first block to
get more distance between the header and the first real text. In their
version (1.4) the distance was 12px due to the `h1:first-child`.
However, this `first-child` thing was not used anymore at all because
of multiple things (wrapper divs inside Root or having breadcrumbs
first). So I removed that.

But I further increased the space from 16 to 20 because I think it looks
better, certainly for realms. Most non-realm pages are not changed by
this as they often have an element with a larger margin-top below the
h1. Some pages are changed, but only slightly and it's still fine.