Fix container size for portrait videos
Previously, the `<InlinePlayer>` would always mirror the aspect ratio
of the video itself, even when its height was greater that its width.
This lead to portrait videos not using the available vertical space
and instead be displayed too small. By swapping the values in these
instances, we make sure that the container will always have a
"landscape" orientation.