Rework configurable colors (remove accent, make happy optional)
This is the result of quite a bit of discussions with various people, 
including two designers. We were not too happy with the previous model.
Now it is simpler: in most cases, there are only 3 instead of 5 
configurable colors. Or even just 2, as most organizations don't change
the grey. This means it's harder to screw things up when configuring
the colors. Previously, 4 colors (where it was implied they should all 
be distinct) needed to be chosen. 

By default, the happy color is the same as the primary color. 

This commit also changes the primary color to the blue that was 
previously the accent color.

Without the corresponding frontend change, this breaks Tobira.

Finally, the way we derive the variants of the configured colors has
changed. Instead of converting to HSL and then changing L, we use the
Lch color space. The L in HSL is not a good indicator for perceived
brightness. In fact, the HSL color model is pretty shitty. So
previously, our modification would have vastly different perceived
brightness changes, depending on the base color. With Lch, this is now
fixed. Well... there is a tiny problem with Lch where brightness
changes for blue-ish colors can introduce some purple tint. But I think
it's very minor and shouldn't be a problem.
There is also the Oklch color space which fixes that problem. Sadly,
in that color space, the L component does not exactly represent the
perceived brightness anymore, so our lighten/darken would need to work
differently.