Add dark mode colors and setup base dark mode logic
The colors are just generated and sent inside `index.html`. The frontend
will switch between the modes purely on the `data-color-scheme`
attribute on the `<html>` tag. Some JS code will set that on load
(hopefully very early, before anything is rendered). Additional logic
with local storage and the like, will be added on top of that.