Add color scheme switching logic and UI (menu in header)
There are two color schemes: "light" and "dark". The active color scheme
is determined by (a) the `prefers-color-scheme` media query and (b) by
a user preference in local storage. The user preference in local storage
can be "light", "dark", "auto" or not-present. The last two cases are
treated equally. We still set "auto" instead of removing the local
storage item, as it might be useful in the future.

If the user preference is set to "light" or "dark" (i.e. something
specific), we use that as scheme. If it is unset or "auto", we check
the `prefers-color-scheme: dark` media query. If it matches, we use
"dark", otherwise we use "light".

As mentioned in the previous commit, the active color scheme is always
stored in the `data-color-scheme` attribute on `<html>`. (If this
attribute does not exist, that's a bug; in that case, we use the light
theme.) This makes it possible for CSS to depend on it (which is done
for the colors). The active color scheme is also stored within a React
context in order for JS code to access it and (more importantly) to get
rerendered when it changes. Note though: only changes through JS lead
to a rerender. If someone just changes the attribute in dev tools, it
does not update the JS. But that's fine, we don't support that.

Some code is directly in `index.html`. That is to make sure the `<html>`
attribute is set as quickly as possible, to avoid a flash of the wrong
color scheme on load. Unfortunately, this duplicates a bit of logic,
but I think it's totally fine.