Change focus behaviour of buttons
This now uses `:focus-visible` instead of `:focus`, to make sure
the focus style won't stick around after clicking the button (which
focuses it). This can still be overwritten be including a `:focus`
selector on the specific button, but it's not possible the other
way around, hence this commit.
Of course this only gets apparent on non-link buttons.