Limit height of description and metadata
On larger screens, the text inside the box uses a minimum of 5 lines,
provided the text is long enough.
The maximum number of lines is dependent on the height of the metadata
box, i.e. if there is a lot of metadata, we can show more text.

On smaller screens however, the maximum number of lines if fixed at 5 lines,
so the user doesn't have to scroll that far to reach the metadata section.

For a similar reason, the number of metadata items to show is limited to 4
items on all screen sizes.

If there is more content to show, both boxes have a button to expand/collapse
the text/metadata.