Fix video overflow for safari
This fixes an issue in safari where the video was overlapping
with buttons and other stuff that should be below the video.

This feels a little heavy handed, but as far as I can tell it
doesn't break anything apart from the upload page, which now
needs an additional `width: 100%` to fill the page horizontally.