Add redirect from video routes to direct video routes if realm does not exist
The idea is to reduce the need for handing out direct links. Before,
"video with page context" links didn't work anymore if the page didn't
exist or didn't include that video. To avoid these breaking links, users
were asked to instead hand out direct links, which wouldn't break but
also didn't have the page context/navigation. But we can just make
the "page context link" do both. In case the page doesn't exist, it just
forwards to the direct link.

In the future we might want to show a note or something, but this is
fine for now.