Make JWT key optional by generating it on the fly if unspecified
This makes setting up Tobira easier as one doesn't have to deal with JWT
at all until later, when one can actually test the uploader and such.
While I first only did that for the initial setup phase, I think using
the generated key is totally fine for most use cases. Yes it changes,
but rotating keys is actually something that's recommended anyway. And
while Opencast has a cache on the JWKS data, if a JWT fails to validate,
it purges the cache and tries again. So restarting Tobira does not cause
any problems.