Fix `import-realm-tree` command
This broke due to the `serde_yaml` update to 0.9. That release changed
how enums are represented in YAML, preferring the appropriate `!Tag`
syntax. This is an improvement, but we don't want to change our huge
test-data file, so we opt-in using the old behavior.

This also fixes the "upload DB dump" GitHub action