Replace most uses of `_react=...` locators
Where possible, text content (like placeholder, label etc) is used
instead. Some elements don't have any text or the text is variable.
In these cases this still uses the `_react=ReactComponent` locator,
but those could also be replaced by using `testId`s on the components
instead. It's not super clear for what reason that would be preferable,
but playwright's docs indicate that it might indeed be (the `_react`
locator is marked as experimental.. well I guess that's clear enough).