Fix "redirect after external login"
We want to redirect freshly-logged-in users to the page they were
coming from. But the previous implementation didn't work. One easy fix
was actually to add `[isLoggedIn]` as dependencies to the `useEffect`
hook. The problem was that the effect was executed a second time after
calling `router.goto` as `router = useRouter()` caused a rerender. And
the second time, the session storage was already cleared and thus,
`router.goto` was called again, but with `"/"`. Users would always be
redirected to the homepage.

Again, this simple specification of `useEffect` dependencies fixed it
in production. But with React's `StrictMode` it was still broken as
the effect would be executed twice. React says that effects+cleanup
should be a noop. So running effect, then cleanup, then effect should
result in the exact same state of the world as only running effect once.
One could fix it by having some boolean local variable, but that seemed
hacky. After reading React docs again, I think triggering this redirect
in an effect just seems like the wrong approach.

So I moved all that logic to the very start of `index.tsx`, before the
initial route is matched. This works well, as users from external
logins would always load that login page "normally" and not via internal
router navigation. I'm not 100% happy with this as now the logic is
even less localized to `Login.tsx`. But I guess its the best option?

Oh and now the login page renders a "you are already logged in" text in
that case. It's not super pretty, but note that users should never be
able to see it as there is no internal-router-link to `/~login` anywhere
on the page if the user is logged in. The user can still type the path
manually, but... oh well, then they see that note :shrug: Seems fine to
me.