Fix `db clear` by not dropping automatically created array types
For each type that we create, postgres adds an additional "array" type
of that type. Trying to drop that array type fails with:

    ERROR: cannot drop type block_type[] because type block_type requires it
    HINT: You can drop type block_type instead.

Deleting the base type also gets rid of the array type. The filter added
in this commit seems to do the trick and should work generally. Related:
https://www.postgresql.org/docs/current/catalog-pg-type.html

Note that this error occurs only occasionally, depending on what order
the type names are returned from our query. In all my tests, the base
types were ordered before the array types, hiding the problem. On a
production instance, they were ordered differently, that's how I
noticed this bug.