Add navigation to user page creation
When the nav is shown (i.e. on larger screens), the
page content is aligned left instead of center to
match the appearence of the other pages. It uses a `max-width`
equal to other text blocks. I think with this, the only remaining
centered text is on the `about Tobira` page.