Remove admin from acl options
The administrator is always allowed to edit/manage
stuff so their inclusion in the acl ui is at
best redundant and at worst misleading for non-admin
users. We still want to show the admin selection entry when
the actual admin is logged in, so in cases where the
acl doesn't explicitly includes it, it gets added manually.