Switch to GitHub LFS for DB dumps
"Our" S3 is currently broken and it might not be maintained anymore.
Using LFS is nice because it's also hosted with GitHub, reducing the
number of points of failure. The dumps are small enough so that we
likely won't run into any limitations.

I already pushed all current DB dumps to the `db-dumps` orphan branch.