Adjust CI scripts to use musl builds on some days (for testing & deploy)
The idea is that that way, we will always test both builds and none will
ever get a bug that remains unnoticed for a long time. Also, annoying
bugs that only appear very rarely or after a long time of running can
be noticed like that as well. Otherwise we wouldn't.