Use `isValidNewOption` for ACL select and use built-in searcher
The problem was: abusing `formatCreateLabel` to check whether creating
a new option is allowed is not something the library expects. If
`formatCreateLabel` returns `false`, then that somehow ends up in the
list of searchable options. For example, `filterOption` is passed
`false` at some point then. Not using `filterOptions` results in the
built-in search never rejecting that option so there is always one
option found. Which has an empty label and is thus shown as a very
narrow div.