Refactor ACL group superset relationship logic
The previous code used `array.sort` which can't really be used that way
since we don't have a total order. This implements topological sort and
while I'm at it, I also changed the `supersetList` function to use the
graph as well. Now it's fairly easy to add configurable superset
relationships.