Improve headings of ACL related UI
On the upload page, everything feeled pretty crammed. So this adds
space. But the HTML tags were also changed. The `h4` for `strong` as
the component appears in multiple contexts, and for some, `h4` is not
correct (doesn't follow `h3`). I also think that `<label>` was not
quite semantically correct. But who knows anything.