Improve handling of `onChange` in ACL selector
Before, we required the second return value of `useState`, i.e. also
the overload where we can pass a function that gets passed the previous
value. That lead to the uploader having to do weird stuff.

Also, the "mutate set and then clone" pattern was used in multiple
places. With this commit, this is simplified as the "make copy" logic
is defined once.