Add focus-style to special remaining elements
Unfortunately, we can't use our focus style everywhere.
There are some special cases where the offset would collide
with `overflow: hidden` and therefore be clipped. So in these
cases, they need a workaround. I had to fiddle around a bit
with paddings and margins, so the enclosed/focused element has some
space between inself and the focus border. Sometimes `outlineOffset`
would work in theory, but then there would be too much space at the
top and bottom and not enough on the sides (subjectively speaking).