Dynamically set player ui width
One does not simply set this player's width.
So this uses a rather roundabout way of doing it.
First, the "scaling" is deduced by comparing the
actual factual video height (`offsetHeight)`, which is
then used to calculate the video width. Why not use the
`offsetWidth` in the first place? Well... setting the width
of something by comparing it to itself can lead to certain
issues.

This also uses a `resizeObserver` to dynamically update the
ui width, because why not.