Raise MeiliSearch requirement to >= 1.4
The main reason for the bump is 1.3 which added ranking details to the
result, which allows us to properly merge the results of multiple
indices. We currently do that in a very hacky way manually. I raised it
to 1.4 as that contains a breaking change regarding filter strings. We
don't yet use filters, but once we add them, it would be nice to not
have to deal with this behavior change between versions.