Use Meili's new ranking score to improve search result merging
This finally gets rid of our hacky previous solution. It's still not
perfect as in theory, comparing scores between indices with different
searchable attributes (and stuff) is not well defined. In practice it's
fine right now. But we should keep an eye on that and fix it at some
point.