Drop support for PG11 (bump minimum required version to 12)
PG11 reaches its EOL in about two weeks anyway. The next Tobira release
contains a bunch of breaking changes already, so it's convenient for
admins to cluster breaking changes into one release. So let's bump it
already.

PG 12 adds a number of features that we might want to take advantage of
in the future.