Add `users` database table
Users don't originate from Tobira, Tobira doesn't manage or own them.
However, this table is still required for a few different things:
- A known list of users with their roles for the ACL selector
- Bookmarks
- Some site settings stored in the account
- Sudo mode timing information

The first point will be implemented in the next commits. All others are
just for the future, but I think adding this information will be easily
be possible later.