Require, store and expose `user_role` field for users (BREAKING)
So far, we had no strict requirements on the user role, we just assumed
there to be exactly one user role in the roles. To make the ACL selector
work properly, we not require this explicitly. This shouldn't be a
problem for anyone as this requirement is likely already fulfilled.