Add `acl` endpoint to `event` API returning roles with labels
Quite a bit of discussion went into how this API should work. There are
various ways to represent ACLs:

1.
    {
      read: [{ role: "ROLE_ANONYMOUS", label: "..." }, ...],
      write: [...],
    }

2.
    {
      read: { "ROLE_ANONYMOUS": "label..." }, ... },
      write: { ... },
    }

3.
    {
      "ROLE_ANONYMOUS": { label: "...", actions: ["read", ...] },
      ...
    }

4.
    [
      { role: "ROLE_ANONYMOUS", label: "...", actions: ["read", ...] },
      ...
    ]

5.
    [
      { role: "ROLE_ANONYMOUS", label: "...", action: "read" },
      { role: "ROLE_ANONYMOUS", label: "...", action: "annotate" },
      ...
    ]

The last one is the way Opencast represents ACLs internally. I dislike that as
it repeats the role and label for each action. Options 1, 2 and 3 are
problematic because of dynamic keys. The actions (read, write, ...) and the
roles itself are used as object keys there. That doesn't work well with GraphQL
which is statically typed. We can make it work by having a custom scalar for
this data structure. But scalars are "atoms" for GraphQL, i.e. one cannot
select a subset of fields then. That's likely not necessary for us but still
maybe not optimal, also because the individual fields are not present in
GraphQL's schema/docs then.

This leaves us with option 4. Only annoying thing here is that we use an array
where each role must only appear once. A map (JSON object) would be fitting to
represent that constraint but oh well.