Fix scrolling footer
The footer was scrolling with certain browser
extensions that add hidden divs to the body.
We were setting all descendant divs of the body
to `100%` height, which caused them to either
push elements up or let users scroll beyond
any visible content. The solution is to only set
the height on the first descendant div.