Replace accent color with grey for button hover background in Paella
We had problems with this as the automatically derived accent colors
(dark and light version) look very different hue-wise. We thought about
just making both configurable but I got an idea that I like better, I
think: don't use the accent color for the hover effect. Instead, a grey
color is used which makes hovering less "flashy", which is good IMO.
The YouTube player for example only has very subtle hover effects. It
does have a changing cursor, which Paella should get as well, but that's
a different issue.