Fix footer once again in presence of browser extensions
Follow up to #1022 where stupidly suggested `:first-of-kind` because
suuurely, browser extensions would only add elements to the end of the
`<body>` but not the beginning, riiight? Of course that's a wrong
assumption. I noticed this with Bitwarden on the login page.

This whole "extensions randomly modifying DOM" thing seems super brittle
to me, oof.