Improve homepage of test deployment
Most of the work was done by uploading and managing videos in Opencast.
This commit changes the blocks of the home page and adds an additional
page, to make everything a bit nicer IMO. The Opencast Summit block
is not really useful on the homepage and makes it very long. It has
been replaced by all Tobira-related talks. The first text block now
explains what Tobira is and links to further resources. The ETHZ test
data thing was shortened as it's really not that important.