Fix router bug accessing `null` history.state`
This lead to a critical error in situation where Tobira redirected
internally after a login. We accidentally replaced the state by `null`,
which was later accessed. Both of these change-snippets solve the bug,
but both are useful. The redirect forward shouldn't just set the state
to null, and the router should never assume the state is non-null.