Filter out past live events from search results in backend
Previously the frontend filtered out those events. But this lead to
weirdness where there "number of hits" were incorrect and if everything
was filtered out, the "no results" was not shown, but no result was
shown either.

This moves this completely to the backend. Which means that due to
communication delay, the frontend could show live events that just
ended. But that's fine I think.

Also, this filtering is not done for the event search in the block edit
UI. Maybe people want to find past live events there :shrug: