Prevent search form submission
Submitting the search `form` currently triggers our CSP, which only
allows form submissions to Studio/the editor, if at all.

That's fine, though, since we don't actually want to ever submit that
form. And even though we don't have a submit button, this can still
happen for example if you press enter in the search input. The `form`
doesn't have an `action`, but the default seems to be to submit to
the current page. 🤷‍♀️