Change title/description/ordering layout
The ordering buttons are now shown right above the
horizontal line as long as there is a description present.
Otherwise it stays at the top right corner of the block.
This needed some fiddling and a few more conditional styles than
I would like, but I think this is the simplest way to achieve the
desired behavior.