Run `cargo test` in CI in release for `master` as well
The assumption was that creating the binary takes forever and thus
even completely rebuilding the debug build for testing is faster. I
tested this locally and it seems not to be the case.
`cargo build --tests --release` after `cargo build --release` is quite
fast. Let's hope this is also true for CI.