Fix series block video items for screen widths 880 - 890
It was somewhat broken for those widths. One constant was also
incorrectly named, which I fixed. I also slightly increased the sizes
for large screens, because it didn't make a ton of sense to me that
the numbers were larger for small screens. All these changes should be
very small though.