Change edit series block label "layout" to "Metadata"
Since we removed the "description above box" layout, there is no point
in calling that "layout" anyway. And now it even clashes with the video
layout option. The name "metadata" didn't convince everyone but I think
it's rather unimportant, so in the interest of time I moved forward with
it. It can still be changed later.