Add logging of useful DB server information
I was in a debugging session yesterday where this kind of information
would have been useful. We were able to install psql to acquire it
ourselves. But (a) having this information without psql and (b) not
having to look up the queries to get that information during debugging
is very useful.