Kill existing DB connections in `db reset/clear`
Existing connections could have cached statements that would lead to
an error, given that everything changed. That is the case for Tobira:
if you run `db reset` while `tobira serve` is running, the latter will
then error on any query. That can be solved by calling
`connection.statement_cache.clear();` at appropriate spots, but that's
not great. Caches have a use. So instead, we just let the `db reset`
kill all connections. Tobira works fine with that as it's using a DB
pool. If connections within the pools get killed, the pool just creates
a new one next time one is requested.