Add container to serve small static media files for UI testing
UI testing also includes testing the Paella player. For that, we don't
want to rely on online servers for the media files. So this commit adds
some small video files, that are served via a dev container.

The files are all managed by git LFS, but since they are on the master
branch, they are downloaded when cloning. But in total they are just 1.2
MB large, so I think that's fine.