Add `relevant_cookies` config to improve upon `relevant_headers`
Once the callback needs to read cookies, specifying `cookie` as relevant
header is likely not great. Usually there are tons of cookies a browser
sends unfortunately. And it's not unlikely for cookies to change with
every request, which would destroy caching. So for the special cookie
header, we have this extra config.