Add realm permissions section
With this, realm editing permissions are initially
restricted to admins. They can add permissions for users
and groups to either `moderate` (i.e. edit content,
add subpages and edit non-critical realm settings) or
`manage` (do everything including deleting realms and
editing their permissions). Permissions are inherited to
children realms, which the ui does not reflect with this
commit. That is added in a following commit.