Add inherited permissions to realm acl selector
This shows the permissions that are inherited from
all ancestor realms as immutable entries in the current
realm's acl selectors. Entries that are already being inherited
can still be added to allow "upgrading" permission
levels, i.e. from "moderate" to "manage". However it is worth
noting that adding an entry with "moderate" level while the
inherited permissions already contain that entry with "manage"
permission, that permissions will not be "downgraded".
Basically the permission level of doubled entries will resolve
to the highest level between both entries.