Add modal for prompting terms-and-conditions consent
This is a really basic consent modal, that is only
closeable by agreeing to the configured terms and
conditions. Users are prompted when first visiting
Tobira.

The consent is saved in local storage, meaning it is
done on a per-device basis rather than per-user.
Users are re-prompted once any texts change in the language they 
originally agreed to.

It is slightly annoying that the logic isn't fully contained in one 
file, but the async nature of the `digest` function is annoying. This
is the easiest way to do this that avoids one-frame flickering of the
popup dialog, for example.