Add modal for prompting t&c consent
This is a really basic consent modal, that is only
closeable by agreeing to the configured terms and
conditions. Users are prompted when first visiting
Tobira.
The consent is saved in local storage, meaning it is
done on a per-device basis rather than per-user.
Users are re-prompted once anything in the T&Cs changes.