Fix navigation to `logout_link` being cancelled
Before, the request to `logout_link` was sent, but cancelled once the
`DELETE` request returned, due to the `window.location.href = "/"`.

But we don't actually want to redirect to `/` if the logout link is set.
That poses a new problem: making sure our `DELETE` request is actually
getting through. Luckily, that's easy with `keepalive`.

One could reduce the amount of code by just using `logout_link ?? "/"`,
but I think the current behavior is good:
- If logout_link is not set, a spinner is shown until the DELETE request
  returned, then the user is redirect.
- If it is set, the user immediately is sent there, with the DELETE
  request being send in the background.