Stop using non-existent Roboto fonts in Paella
The Paella skin we use loads Robot files and sets Roboto as font for
the whole player. For some time now, those requests to the font files
failed, creating annoying errors. So unless you already had Robot
installed locally,  Paella used the font set by Tobira.

We discussed this internally and decided that having two similar,
but still different fonts on a website is not a good idea. So this
removes the font declarations in the skin. In a ... hacky way sadly.
But we already heavily modify Paella code to our advantage.